QFW: Using the DVM causes application failure
Using mpirun or prun to spin up the QFw infrastructure seems to lead
to a situation where mpirun calls to run the application eventually
fail with 195 return code. After debugging it appears like:
prun_common.c:defhandler() is being called setting the status to -61
	-61 = PMIX_ERR_LOST_CONNECTION
	Eventually: prun_common() fails when it checks the status here:
	/* if we lost connection to the server, then we are done */
	(PMIX_ERR_LOST_CONNECTION == rc || PMIX_ERR_UNREACH == rc) {
		print_current_time(date);
		fprintf(logging, "%s:%s:%d:%d\n", date, __FILE__, __LINE__, rc);
		goto DONE;
	}
To avoid this issue use defw_exec_remote_cmd() to startup the
infrastructure processes. This is done via paramiko.

Since the QPM/QRC uses the PRTE DVM to run the applications and the PRTE
DVM uses srun the SLINGSHOT_VNIS are set up properly and everything
works.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>