QFw: Support python virtual environments
Added support for python virtual environments. Only create a virtual
environment if one is not specified. User script should set
QFW_VENV_PATH to the venv directory. The scripts will change the python
symlinks to the defwp. They will change them back in the teardown stage.
If something goes wrong in the middle of the test qfw_restore_venv.sh
can be called explicitly to restore the venv to its previous state.

The steps are now:
- Create a venv
- Activate the venv
- Run the script
   - call qfw_setup.py
   - call qfw_srun.py
   - call qfw_teardown.py

Signed-off-by: Amir Shehata <shehataa@ornl.gov>